.claude/skills/sketch-dev/scripts/plugins/pyproject.toml
[project]
name = "sketch-plugin-tracker"
version = "0.1.0"
description = "Track, categorize, and monitor Sketch plugins from the official extensions page"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "crawl4ai>=0.4.0",
    "pydantic>=2.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[project.scripts]
sketch-plugins = "pipeline:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["lib"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
